requirements.txt
asgiref==3.2.7
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.7
firebase-admin==4.3.0
gcloud==0.17.0
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
httplib2==0.18.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jws==0.1.3
msgpack==1.0.0
oauth2client==3.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
zope.interface==5.1.0
